requirements.txt
# FastAPI and related dependencies
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic==2.5.2

# Mistral AI client (assuming it's a custom package, replace with the actual package name if different)
mistralai==1.0.0

# FAISS for similarity search
faiss-cpu==1.7.3

# PyPDF2 for PDF text extraction
PyPDF2==3.0.1

# Django and related dependencies
Django==5.1.4

# Requests for making HTTP requests
requests==2.31.0

# Logging
logging==0.5.1.2
